Removed commented code.
For the record, this branch adds a missing call to flush in the use of CharsetEncoder. While this may seem
benign, this missing call was triggering a nasty bug when using some specific JDK, namely JDK17.0.7+7, JDK18, JDK20, but not JDK8/11 or 21-ea+25-2212.

That bug led to SipHash hashes being incorrectly computed. The reason for this being that the buffer passed to SipHash is not correctly filled with label values by the code in method labelsId.

Inspection of the Java code, apart from the missing call to CharsetEncoder.flush, does not show any reason for this issue. We could not isolate a reproducible test case but we had a Warp 10 instance archive from a user which could be used for reproducing the issue.

Code modification were inserted to display the buffer content passed to the SipHash computation and the issue was identified.

As the flush method in CharsetEncoder.flush does not modify the output ByteBuffer passed as parameter when using the UTF_8$Encoder we suspected the issue to be related to the JIT which may produce code which leaves the buffer in an incoherent state. Adding a call to the flush method adds a call site in the byte code for a method which modifies the state of the CharsetEncoder instance (setting the state to ST_FLUSHED), this is the reason why adding that call probably solved the issue as the behavior of the JIT is changed.

Tests done with the JDKs which experienced the issue but with the JIT disabled (-Djava.compiler=NONE) or with inlining disabled (-XX:FreqInlineSize=1 -XX:MaxInlineSize=1 -XX:MaxTrivialSize=1) showed that the problem did not occur when either of those configurations were used.

The conclusion is therefore that the missing calls to flush had a side effect of triggering a bug present in the JIT implementation in some JDKs. Unfortunately we were unable to craft a simple enough test case to be able to open an issue. Since the issue does not seem to be present in JDK 21-ea, we can hope that changes were done to the JIT which solved the issue.